Replace all C++ typedefs with 'using'
It is clearer and easier to read.

c.f.
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rt-using

+++ src/rpc/core_rpc_server.cpp
@@ -2100,7 +2100,7 @@ namespace cryptonote { namespace rpc {
     uint8_t version = req.version > 0 ? req.version : blockchain.get_next_hard_fork_version();
     res.version = blockchain.get_current_hard_fork_version();
     res.enabled = blockchain.get_hard_fork_voting_info(version, res.window, res.votes, res.threshold, res.earliest_height, res.voting);
-    res.state = blockchain.get_hard_fork_state();
+    res.state = static_cast<uint32_t>(blockchain.get_hard_fork_state());
     res.status = STATUS_OK;
     return res;
   }
